requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
